+++ src/libcmd/include/nix/cmd/meson.build
@@ -20,4 +20,5 @@ headers = files(
   'network-proxy.hh',
   'repl-interacter.hh',
   'repl.hh',
+  'unix-socket-server.hh',
 )
